add Assert context for running database tests in a transaction
This adds a `DbTests` Assert context that has an `around` block
for running tests in a transaction and rolling them back after the
tests have run.

This allows not only speeds up your tests b/c you are not actually
persisting any data, it also allows you to do parallel test runs
b/c the database interactions are now multi-process safe because
of the transactions.

All of the ActiveRecord interactions are behind the adapter so
this just tests the adapter API.  In this case, I doubt the
ActiveRecord interactions will change over major versions (ie the
`transaction` and `rollback` apis are pretty stable), a design
principal is this gem cannot directly require ActiveRecord.
Therefore, all interactions must be behind the adapters.